+++ src/core/hashtable.h
@@ -46,7 +46,7 @@ IGRAPH_PRIVATE_EXPORT igraph_error_t igraph_hashtable_addset(igraph_hashtable_t 
                                                              const char *elem);
 IGRAPH_PRIVATE_EXPORT igraph_error_t igraph_hashtable_addset2(igraph_hashtable_t *ht,
                                                               const char *key, const char *def,
-                                                              const char *elem, int elemlen);
+                                                              const char *elem, size_t elemlen);
 IGRAPH_PRIVATE_EXPORT igraph_error_t igraph_hashtable_get(igraph_hashtable_t *ht,
                                                               const char *key, char **elem);
 IGRAPH_PRIVATE_EXPORT igraph_error_t igraph_hashtable_getkeys(igraph_hashtable_t *ht,
